apps/api/requirements.txt
# --- Web framework ---
fastapi==0.115.0
uvicorn[standard]==0.30.6

# --- Database ---
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.3
pgvector==0.3.6

# --- Cache ---
redis==5.0.8

# --- Validation / Settings ---
pydantic==2.9.2
pydantic-settings==2.5.2
email-validator==2.2.0

# --- Auth (Day 2+) ---
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.10

# --- Utilities ---
python-dotenv==1.0.1
httpx==0.27.2
tenacity==9.0.0
aiosmtplib==3.0.2

# --- LLM + Embeddings (Day 6, 8) ---
# Falls back to mock provider when API keys are missing.
anthropic==0.39.0
openai==1.51.0
tiktoken==0.8.0

# --- Background Jobs (v0.6) ---
arq==0.26.1

# --- Structured Logging (v0.6) ---
structlog==24.4.0

# --- Observability (v0.6) ---
opentelemetry-api==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-instrumentation-sqlalchemy==0.48b0
opentelemetry-instrumentation-redis==0.48b0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
sentry-sdk[fastapi]==2.14.0

# --- Dev ---
black==24.8.0
ruff==0.6.9
pytest==8.3.3
pytest-asyncio==0.24.0
